+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - when pasting html from the clipboard as markdown null characters and Windows
   line breaks will now be removed as well as all tags will be interpreted over
   multiple lines (for [Issue #645](https://github.com/pbek/QOwnNotes/issues/645))
+- added more Catalan translation (thank you MGuerra)
 
 ## 17.07.9
 - fixed a possible crash when using scripts with custom actions in the note